requirement.txt
finlab
streamlit
plotly
PyYaml
openai
lxml
IPython
stqdm
opencc
streamlit-authenticator